Revision to include Sommerfeld T-dependent potentials.
These include the effect of excitation of the electronic degrees of
freedom in the potential and operate only with generic.atvf potentials
and only if there is a file "sommerfeld.in" in the directory where the code is
running.

The temperature variable is now available to the dynamics file and generic_atvf
This causes some messy circular dependencies of mod files, which should be 
removed with a reengineering of the code to merge _magnetic _sommerfeld and EAM schemes.

Also, a few bugfixes to catch rare divide by zero errors for sqrt(eps)=0.

Integration of the Magnetic potentials is still flaky.  Magnetic moments 
are evolved dynamically rather than iterated to the Born-Oppenheimer 
surface, this makes the forces weakly history dependent which messes up the 
conjugate gradients quench.